# src/main/java/backend/hobbiebackend/model/jwt/JwtResponse.java
package backend.hobbiebackend.model.jwt;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class JwtResponse {
    private String accessToken;
}